fix: start menu quitting
Qurrently when pressing `q` or `<Ctrl>-C` the program throws an exception.
Reason for this is that on those two cases `terminal_menu.show()` returns `None` instead of a valid index.

Therefore I decided to add a simple `Bye` message and then exit. Ideas for improve this handling are welcome :)