fix: handle Request.set_api() errors
Added error handling for when env variables are missing.
I decided to just exit the program, since without those it cannot run
anyways.
Not fully satisfied with how the missing message gets created, but yea.

Also added error handling for when `client_id`, or `client_secret` are
invalid, which causes the `api.fetch_token()` to throw an exception.