refactor: use a game loop for `Interface`
Currently `Interface` does not use a "game loop" but rather recursive
functions.
This might cause problems for power users, but also adds a layer of
complexity.
Therefore I have changed it to use a simple `while (True)` loop.

Further more I started a refactor of `Interface` in a way which makes it reusable
throughout the entire codebase.