feat: add campus prompt function
Currently this is just a hardcoded list. Maybe we can make it dynamic
later, but then it's an additional API call which slows down the
program.
Maybe we could make the API call in a lazy way, so only all campuses are
requested, once they are actually needed, and then stored.
But then again it's not like a new campus comes every day.