fix: `OddsOfFailing` bonus marks
Currently the `OddsOfFailing` does not take bonus marks into account, so
makrs which can go over `100`, leading to incorrect results.
The easiest way to takle this issue is to cap the mark just to `100`.
But maybe we can come up with a more sophisticated formula which takes
bonus points into account too 🤔.