feat: improve loading animation system
Added a cool more colorful dot spinner thingy.
Also made the entire thing work as function decorators and as a
context manager.

Further more I've added a `status_message` feature, which is useful,
when for example a rate limit is exceeded. Instead of printing it
directly to the console now the spinner directly can display it in a
good manner.

Before implemening it on my own I have looked at existing Python
modules. But unfortunately they are either not really maintained
anymore, or don't really provide the features I want lol.